(feat): add vite and refactor javascript into different snippets
