+++ resources/js/index.js
@@ -1,0 +1,2 @@
+import './snippets/styles.js';
+import './snippets/uploadEventHandlers.js';
